refactor(organisms/ArticleImageCarousel): molecule -> organism으로 분류 변경
